Organize origin types
