+++ src/Origins/Reviewers/ConstantReviewer.php
@@ -5,11 +5,11 @@ declare(strict_types=1);
 namespace PhpCfdi\SatCatalogosPopulate\Origins\Reviewers;
 
 use LogicException;
-use PhpCfdi\SatCatalogosPopulate\Origins\ConstantOrigin;
 use PhpCfdi\SatCatalogosPopulate\Origins\OriginInterface;
 use PhpCfdi\SatCatalogosPopulate\Origins\ResourcesGatewayInterface;
 use PhpCfdi\SatCatalogosPopulate\Origins\Review;
 use PhpCfdi\SatCatalogosPopulate\Origins\ReviewStatus;
+use PhpCfdi\SatCatalogosPopulate\Origins\Types\ConstantOrigin;
 
 final readonly class ConstantReviewer implements ReviewerInterface
 {
