+++ src/Origins/Reviewers/Nomina12eReviewer.php
@@ -8,11 +8,11 @@ use Exception;
 use HeadlessChromium\BrowserFactory;
 use HeadlessChromium\Dom\Selector\XPathSelector;
 use LogicException;
-use PhpCfdi\SatCatalogosPopulate\Origins\ConstantOrigin;
-use PhpCfdi\SatCatalogosPopulate\Origins\Nomina12eOrigin;
 use PhpCfdi\SatCatalogosPopulate\Origins\OriginInterface;
 use PhpCfdi\SatCatalogosPopulate\Origins\ResourcesGatewayInterface;
 use PhpCfdi\SatCatalogosPopulate\Origins\Review;
+use PhpCfdi\SatCatalogosPopulate\Origins\Types\ConstantOrigin;
+use PhpCfdi\SatCatalogosPopulate\Origins\Types\Nomina12eOrigin;
 use RuntimeException;
 
 final class Nomina12eReviewer implements ReviewerInterface
